fixed useauth
