+++ client/src/components/Authpage/Signin.tsx
@@ -13,7 +13,7 @@ export default function Signin() {
    });
    const verifyLogin = () => {
       axios
-         .post('http://localhost:8080/auth/signin', {
+         .post('http://localhost:8080/api/auth/signin', {
             params: {
                userName: values.userName,
                password: values.password,
