+++ client/src/components/personal-profile/FollowerSearchBar/followerSearchBar.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect }  from 'react';
-import { SearchItem } from '../followerSearchBar/searchItem/searchItem';
+import { SearchItem } from './searchItem/searchItem';
 import axios from 'axios';
 import { TextField } from '@mui/material';
