+++ global.d.ts
@@ -7,3 +7,4 @@ declare module '*.jpg' {
    const content: any;
    export default content;
 }
+
