+++ server/database/routes/index.ts
@@ -90,7 +90,7 @@ router.put('/user/addFollowed', (req: Request, res: Response) => {
 
 router.put('/user/removeFollowed', (req: Request, res: Response) => {
   return controllers.updateUser(req.body.userName, 'followingList', req.body.value)
-    .then((results) => {
+    .then((results:any) => {
       console.log(`/user/removeFollowed: Success removing ${req.body.value} from following list`, results);
       res.status(204).send(results);
       res.end();
@@ -101,6 +101,7 @@ router.put('/user/removeFollowed', (req: Request, res: Response) => {
       res.end();
     })
 });
+
 
 //TODO: update owned services list
 router.put('/user/services', async (req: Request, res: Response) => {
