+++ types.ts
@@ -2,10 +2,11 @@ import session from 'express-session';
 declare module 'express-session' {
    export interface SessionData {
       username: string;
+      user: string;
       passport: {
          user: {
             id: string;
-            username: string;
+            user: string;
          };
       };
    }
